requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
eyeD3==0.8
feedparser==5.2.1
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
peewee==3.0.17
Pillow==5.0.0
python-dateutil==2.6.1
pytimeparse==1.1.7
pytz==2018.3
requests==2.18.4
six==1.11.0
Unidecode==1.0.22
urllib3==1.22
Werkzeug==0.14.1
